adjust swiftlint `attributes` rule
why? the rule enforces that attributes (eg `@MainActor`, `@Model`, `@Query`, `@ViewBuilder` etc) must be on the same line as the definition they're attached to for variables, and on a separate line for types and functions).
The old configuration also enforced that attributes with parameters always had to be on the same line as the associated variable definition.

This is not particularly good, since in cases where there will be a large number of arguments, having to also tack the variable name and type at the end of the line makes it visually difficult to read the resulting code (plus also you'll eventually run into the line length limit.

Signed-off-by: Lukas Kollmer <hey@lukaskollmer.de>